Done some renamings
