+++ problemsets/problemset-06/CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 3.12)
 # Define all source files
 set(TASKS_SRCS  tasks/LastInFirstOut.cpp
                 tasks/ReversePolishNotation/task-61-postfix-notation.cpp
-        tasks/BracketCounting/task-62-counting-brackets.cpp)
+                tasks/BracketCounter/task-62-counting-brackets.cpp)
 
 # Define all tests
 set(TESTS   utests/utest-task-61.cpp
