+++ problemsets/problemset-06/tasks/BracketCounter/task-62-counting-brackets.cpp
